+++ l10n/fr.php
@@ -1,0 +1,24 @@
+<?php
+$TRANSLATIONS = [
+    'Signing in…' => 'Connexion en cours…',
+    'Signing you in, please wait…' => 'Connexion en cours, veuillez patienter…',
+    'Continue' => 'Continuer',
+    'SAML Provider' => 'Fournisseur SAML',
+    'Entity ID must not be empty' => 'L\'identifiant d\'entité ne doit pas être vide',
+    'ACS URL must be a valid URL' => 'L\'URL ACS doit être une URL valide',
+    'ACS URL must use http:// or https://' => 'L\'URL ACS doit utiliser http:// ou https://',
+    'SAML Provider runs on HTTPS; cleartext HTTP is not allowed for service providers in production.' => 'Le fournisseur SAML fonctionne avec HTTPS ; HTTP non chiffré n’est pas autorisé en production.',
+    'Unsupported NameID format' => 'Format NameID non pris en charge',
+    'Attribute mapping must be valid JSON' => 'Le mappage d\'attributs doit être un JSON valide',
+    'SP certificate is not a valid X.509 PEM certificate' => 'Le certificat SP n’est pas un certificat X.509 PEM valide',
+    'Requiring signed requests needs the SP certificate' => 'Les demandes signées exigent le certificat SP',
+    'A service provider with this Entity ID already exists' => 'Un fournisseur de service avec cet ID d’entité existe déjà',
+    'Continue to {service}' => 'Continuer vers {service}',
+    'You are about to sign in to this connected service with your Nextcloud account.' => 'Vous allez vous connecter à ce service avec votre compte Nextcloud.',
+    'Not found' => 'Introuvable',
+    'SLO URL must be a valid http:// or https:// URL' => 'L’URL SLO doit être une URL http:// ou https:// valide',
+    'Attribute mapping must be a JSON object' => 'Le mappage d’attributs doit être un objet JSON',
+    'Attribute mapping values must be uid, displayName, or mail' => 'Les valeurs du mappage doivent être uid, displayName ou mail',
+    'Service provider not found' => 'Fournisseur de service introuvable',
+];
+$PLURAL_FORMS = "nplurals=2; plural=(n != 1);";